Refactor extract LoginErrorBar to 'gui.auth.sign_in' namespace
