+++ securedrop_client/gui/auth/__init__.py
@@ -17,4 +17,4 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 # Import classes here to make possible to import them from securedrop_client.gui.auth
-from securedrop_client.gui.auth.sign_in import SignInButton  # noqa: F401
+from securedrop_client.gui.auth.sign_in import LoginErrorBar, SignInButton  # noqa: F401
