+++ securedrop_client/gui/auth/sign_in/__init__.py
@@ -18,3 +18,4 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 # Import classes here to make possible to import them from securedrop_client.gui.auth.sign_in
 from securedrop_client.gui.auth.sign_in.button import SignInButton  # noqa: F401
+from securedrop_client.gui.auth.sign_in.error_bar import LoginErrorBar  # noqa: F401
